move p9_write_file to p9fs.rs
